Fix potential actual words within the board
While intially thought to be rare, in actuality the experimental
likelyhood of having a board containing an actual word is approximately
12.7715% over a million trials (I could not figure out how to find the
exact value). The code used for that experiment can be found here:
https://gist.github.com/6167656e74323431/17aa0db00b7c0fe935edf611c04a5880

This new method uses a trie constructed from the word list, and checks
starting at each letter for a word. With a trie the complexity is
~O(n^2) amortized for the whole board rather than ~O(n^3) amortized.
Noting that the O(n^2) might be run multiple times since breaking one
word might simulateously fix another.